refactor: rename authRoutes to authRoute
