+++ src/routes/authRoute.ts
